REFACTORING

+++ include/constants.hpp
@@ -3,8 +3,8 @@
 namespace kmeans {
 	const int VIDEO_WIDTH = 640;
 	const int VIDEO_HEIGHT = 480;
-	const int COLOR_SCALE = 1.0f;
-	const int SPATIAL_SCALE = 0.5f;
+	const float COLOR_SCALE = 1.0f;
+	const float SPATIAL_SCALE = 0.5f;
 	const int K_MIN = 2;
 	const int K_MAX = 12;
 	const int SAMPLE_COUNT = 2000;
